Refactoring the repository and Unit tests
