+++ Interview/Interview.Test/RepositoryTest.cs
@@ -7,14 +7,15 @@ namespace Interview.Test
     public class RepositoryTest
     {
         
+        /// <summary>
+        /// Test method to  GetAll
+        /// </summary>
        
         [Fact]
         public void GetAllStoreableRecordsFromRepository()
         {
-
-
-           
-            IRepository<IStoreable<int>, int> repository = new Repository<IStoreable<int>, int>();
+            
+            IRepository<IStoreable<int>, int> repository =  new MemoryRepository<IStoreable<int>, int>();
             repository.Save(new Storeable<int> { Id = 1 });
             repository.Save(new Storeable<int> { Id = 2 });
             repository.Save(new Storeable<int> { Id = 3 });
@@ -23,15 +24,18 @@ namespace Interview.Test
 
             Assert.Collection(repository.GetAll(), item => item.Id = 1, item => item.Id = 2, item => item.Id = 3, item => item.Id = 4)
 ;
-
-
+            
         }
+
+        /// <summary>
+        /// Test method to Get record by Id
+        /// </summary>
 
         [Fact]
         public void GetStoreableRecordFromRepositoryById()
         {
             
-            IRepository<IStoreable<int>, int> repository = new Repository<IStoreable<int>, int>();
+            IRepository<IStoreable<int>, int> repository = new MemoryRepository<IStoreable<int>, int>();
             repository.Save(new Storeable<int> { Id = 1 });
             repository.Save(new Storeable<int> { Id = 2 });
             repository.Save(new Storeable<int> { Id = 3 });
@@ -41,15 +45,17 @@ namespace Interview.Test
 
             Assert.Equal(expectedResult.Id, repository.Get(2).Id);
 
-
-
+            
         }
+
+        /// <summary>
+        /// Test method of Delete form repository
+        /// </summary>
         [Fact]
         public void DeleteStoreableRecordFromRepositoryById()
         {
             
-
-            IRepository<IStoreable<int>, int> repository = new Repository<IStoreable<int>, int>();
+            IRepository<IStoreable<int>, int> repository = new MemoryRepository<IStoreable<int>, int>();
             repository.Save(new Storeable<int> { Id = 1 });
             repository.Save(new Storeable<int> { Id = 2 });
             repository.Save(new Storeable<int> { Id = 3 });
@@ -59,26 +65,25 @@ namespace Interview.Test
             repository.Delete(2);
 
             Assert.Null(repository.Get(2));
-
-
+            
 
         }
-
+        /// <summary>
+        /// Test method to Save repository 
+        /// </summary>
         [Fact]
         public void SaveStoreableRecordInRepository()
         {
-
-
+            
             IStoreable<int> exptectedResult = new Storeable<int>() { Id = 1 };
             
-            IRepository<IStoreable<int>, int> repository = new Repository<IStoreable<int>, int>();
+            IRepository<IStoreable<int>, int> repository =  new MemoryRepository<IStoreable<int>, int>();
 
             repository.Save(exptectedResult);
 
             Assert.Equal(exptectedResult,repository.Get(1));
 
-
-
+            
         }
     }
 }
